Use generic BatchActions component
